fix: outdated test `test_compare_against_str`
I don't know if this is the actually expected behaviour but at least the test passes now. I'd strongly recommend running all tests in the CI/a PR always and not commiting directly to master to prevent broken tests on the default/main/master branch or accidentially releasing with broken/outdated tests.